Add nil to maps
Because refllectwalk always indirects pointer values, copystructure
can't determine the type of a nil pointer. This isn't an issue with
struct fields, because the field zero value will already be the properly
typed nil, and slices are already initialized with the correct capacity
and zero values. When inserting nil into a map however, a zero value
reflect.Value indicates that the key be removed bu SetMapIndex.

This patch creates a properly typed nil value to insert into a map.