Don't call Interface() on a non-pointer value
Calling Interface() on a relect.Value requires a copy, so it's not safe
to call on a non-pointer value to check for the lock methods. This is OK
since a pointer to a sync.Locker still satisfies the interface. It will
however require that cal call to Copy with Lock set to true can only
accept a pointer argument, even if the value is safe to Lock/Unlock.