copy interface pointers (arbitrarily deep)
Fixes #16

This properly converts a reflected value to `*interface{}` (with any
number of `*` nesting) so that it is settable.

When a field is `interface{}` (no pointer), any value is settable to it.
But when a field is `*interface{}` (one or more pointers), then pointer
values are not assignable to it, it must _exactly_ match `*interface{}`.

The code added in this PR checks for this scenario and converts.