Add ShallowCopiers for shallow copying pointer values
This adds the ShallowCopiers field, which allows specifying a set of
pointer types that, when encountered during copy, will be shallow copied
and not walked into.

This functionality is being introduced to address missing coverage in
the current shallow copy and custom copier functionality in that it only
works off of struct fields and concrete values, specifically. This means
that it's not possible to shallow copy off of individual map fields,
etc.

ShallowCopiers works by hooking into the PointerValue interface in
reflectwalk, and if present in the ShallowCopiers type set, treats the
value as a primitive and does not enter the pointer.

Note that non-pointer value types in ShallowCopiers should have no
effect, although they are not checked, so their behavior is undefined.